refactor(hooks/builtins): inline GetEnvironmentInfo and isGitRepo
session.GetEnvironmentInfo and session.isGitRepo had a single production caller (the add_environment_info builtin) and no callers elsewhere. Move both — along with their helpers (boolToYesNo, getOperatingSystem, getArchitecture) — into pkg/hooks/builtins where they're used, dropping the cross-package import. The function is privatized in the process; tests move with it as internal-package tests so they can keep exercising the helpers.
