+++ pkg/hooks/builtins/add_environment_info_test.go
@@ -1,4 +1,4 @@
-package session
+package builtins
 
 import (
 	"os"
@@ -51,7 +51,7 @@ func TestGetEnvironmentInfo(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			dir := tt.setupFunc()
-			info := GetEnvironmentInfo(dir)
+			info := getEnvironmentInfo(dir)
 
 			gitStatus := "No"
 			if tt.expectGit {
@@ -80,7 +80,7 @@ func TestGetEnvironmentInfoIntegration(t *testing.T) {
 	wd, err := os.Getwd()
 	require.NoError(t, err)
 
-	info := GetEnvironmentInfo(wd)
+	info := getEnvironmentInfo(wd)
 
 	assert.Contains(t, info, "Here is useful information about the environment you are running in:")
 	assert.Contains(t, info, "<env>")
