+++ pkg/hooks/builtins/git_test.go
@@ -1,4 +1,4 @@
-package session
+package builtins
 
 import (
 	"os"
